Move all preamble stuff back into one file
